Refine shared UI ownership and module boundaries
