+++ ui/coreShared/ts/tests/testUtils/domTestLifecycle.ts
@@ -8,17 +8,28 @@ type DomTestLifecycleOptions = {
 
 export function installDomTestLifecycle(options: DomTestLifecycleOptions = {}) {
 	let restoreDomEnvironment: (() => void) | undefined
+	const renderedCleanups: Array<() => Promise<void> | void> = []
 
 	beforeEach(() => {
+		renderedCleanups.length = 0
 		restoreDomEnvironment = installDomEnvironment(options.url).cleanup
 	})
 
 	afterEach(async () => {
 		try {
+			for (const cleanup of renderedCleanups.reverse()) await cleanup()
 			await options.afterTest?.()
 		} finally {
+			renderedCleanups.length = 0
 			restoreDomEnvironment?.()
 			restoreDomEnvironment = undefined
 		}
 	})
+
+	return {
+		trackRendered<T extends { cleanup: () => Promise<void> | void }>(rendered: T): T {
+			renderedCleanups.push(rendered.cleanup)
+			return rendered
+		},
+	}
 }
