+++ ui/statoblast/ts/features/security-pools/components/SecurityVaultSection.tsx
@@ -4,9 +4,7 @@ import * as transactionReviewCopy from '@zoltar/ui-core-shared/copy/transactionR
 import { useEffect, useId, useRef, useState } from 'preact/hooks'
 import { AddressValue } from '@zoltar/ui-core-shared/components/AddressValue.js'
 import { ActionLauncherCard } from '@zoltar/ui-core-shared/components/ActionLauncherCard.js'
-import { Badge } from '@zoltar/ui-core-shared/components/Badge.js'
 import { CurrencyValue } from '@zoltar/ui-core-shared/components/CurrencyValue.js'
-import { EntityCard } from '@zoltar/ui-core-shared/components/EntityCard.js'
 import { ErrorNotice } from '@zoltar/ui-core-shared/components/ErrorNotice.js'
 import { FormInput } from '@zoltar/ui-core-shared/components/FormInput.js'
 import { LookupFieldRow } from '@zoltar/ui-core-shared/components/LookupFieldRow.js'
@@ -21,8 +19,6 @@ import { TimestampValue } from '@zoltar/ui-core-shared/components/TimestampValue
 import { TokenApprovalControl } from '@zoltar/ui-core-shared/components/TokenApprovalControl.js'
 import { TransactionActionButton } from '@zoltar/ui-core-shared/components/TransactionActionButton.js'
 import { TransactionNetworkValue } from '@zoltar/ui-core-shared/components/TransactionNetworkValue.js'
-import { VaultMetricGrid } from './VaultMetricGrid.js'
-import { WarningSurface } from '@zoltar/ui-core-shared/components/WarningSurface.js'
 import { normalizeAddress, sameAddress } from '@zoltar/ui-core-shared/lib/address.js'
 import { formatCurrencyBalance, formatCurrencyInputBalance, formatDuration } from '@zoltar/ui-core-shared/lib/formatters.js'
 import { balanceShortage } from '@zoltar/ui-core-shared/lib/inputs.js'
@@ -45,236 +41,16 @@ import {
 	getSecurityVaultWithdrawableRepAmount,
 	getSelectedVaultOwner,
 	hasValidSecurityVaultOraclePrice,
-	isOracleManagerPriceUsable,
 	isSecurityVaultDepositBelowMinimum,
 	isSelectedVaultOwnedByAccount as isSelectedVaultOwnedByAccountHelper,
 	MIN_SECURITY_VAULT_REP_DEPOSIT_ATTO_REP,
 } from '../lib/securityVault.js'
-import type { StagedOracleOperation } from '@zoltar/ui-core-shared/types/contracts.js'
 import type { ReadinessAction, SecurityVaultSectionProps } from '../../types.js'
-type SelectedVaultSummarySectionProps = Pick<SecurityVaultSectionProps, 'repPerEthPrice' | 'repPerEthSource' | 'repPerEthSourceUrl' | 'selectedPoolStatoblastSecurityMultiplierBps'> & {
-	capacityOwnershipAttoRep: bigint
-	currentVaultIsHealthy?: boolean | undefined
-	securityVaultDetails: NonNullable<SecurityVaultSectionProps['securityVaultDetails']>
-	selectedVaultIsOwnedByAccount: boolean
-	variant?: 'embedded' | 'record'
-}
-type VaultActionModal = 'claim-fees' | 'deposit-rep' | 'withdraw-rep' | undefined
-type QueuedVaultOperationStatus = 'executed' | 'failed' | 'manual-queued' | 'missing' | 'queued' | 'refreshing' | undefined
-type QueuedVaultOperationView = {
-	amount: bigint | undefined
-	isPendingSlot: boolean
-	operationId: bigint
-}
-export function SelectedVaultSummarySection({ repPerEthPrice, repPerEthSource, repPerEthSourceUrl, capacityOwnershipAttoRep, currentVaultIsHealthy, securityVaultDetails, selectedPoolStatoblastSecurityMultiplierBps, selectedVaultIsOwnedByAccount, variant = 'record' }: SelectedVaultSummarySectionProps) {
-	const summaryTitle = <span>{securityPoolCopy.vaultSummary}</span>
+import { SelectedVaultSummarySection } from './SelectedVaultSummarySection.js'
+import { getQueuedVaultOperation, getQueuedVaultOperationStatus, VaultQueuedOperationStatusCard } from './VaultQueuedOperationStatusCard.js'
 
-	const embeddedContent = (
-		<div className='security-pool-selected-vault-summary security-pool-browse-vault-list'>
-			<div className='security-pool-browse-vault-row'>
-				<div className={`security-pool-browse-vault-row-top security-pool-browse-vault-row-top-compact${securityVaultDetails.badDebtAttoEth > 0n ? ' with-bad-debt' : ''}`}>
-					<div className='security-pool-browse-vault-row-title'>
-						<div className='security-pool-browse-vault-row-id'>
-							<strong>
-								<AddressValue address={securityVaultDetails.vaultAddress} />
-							</strong>
-						</div>
-					</div>
-					<div className='security-pool-browse-vault-row-kpi'>
-						<span>{securityPoolCopy.currentCapacityOwnershipAttoRep}</span>
-						<strong>
-							<CurrencyValue value={capacityOwnershipAttoRep} suffix={commonCopy.rep} />
-						</strong>
-					</div>
-					<div className='security-pool-browse-vault-row-kpi'>
-						<span>{commonCopy.poolHeldVaultRepBackingAttoRep}</span>
-						<strong>
-							<CurrencyValue value={securityVaultDetails.vaultAttoRepBacking} suffix={commonCopy.rep} />
-						</strong>
-					</div>
-					<div className='security-pool-browse-vault-row-kpi'>
-						<span>{commonCopy.disputeStakedAttoRep}</span>
-						<strong>
-							<CurrencyValue value={securityVaultDetails.disputeStakedAttoRep} suffix={commonCopy.rep} />
-						</strong>
-					</div>
-					{securityVaultDetails.badDebtAttoEth > 0n ? (
-						<div className='security-pool-browse-vault-row-kpi'>
-							<span>{securityPoolCopy.badDebt}</span>
-							<strong>
-								<CurrencyValue value={securityVaultDetails.badDebtAttoEth} suffix={commonCopy.eth} />
-							</strong>
-						</div>
-					) : null}
-				</div>
-			</div>
-		</div>
-	)
-	const gridContent = (
-		<VaultMetricGrid
-			associatedRepPerCapacityBps={securityVaultDetails.associatedRepPerCapacityBps}
-			badDebtAttoEth={securityVaultDetails.badDebtAttoEth}
-			layout='grid'
-			disputeStakedAttoRep={securityVaultDetails.disputeStakedAttoRep}
-			isCurrentlyHealthy={currentVaultIsHealthy}
-			poolHeldRepPerCapacityBps={securityVaultDetails.poolHeldRepPerCapacityBps}
-			vaultAttoRepBacking={securityVaultDetails.vaultAttoRepBacking}
-			repPerEthPrice={repPerEthPrice}
-			repPerEthSource={repPerEthSource}
-			repPerEthSourceUrl={repPerEthSourceUrl}
-			selectedPoolStatoblastSecurityMultiplierBps={selectedPoolStatoblastSecurityMultiplierBps}
-			capacityOwnershipAttoRep={capacityOwnershipAttoRep}
-			claimableFeesAttoEth={securityVaultDetails.claimableFeesAttoEth}
-		/>
-	)
-	if (variant === 'embedded')
-		return (
-			<SectionBlock density='compact' headingLevel={4} title={summaryTitle} variant='embedded'>
-				{embeddedContent}
-			</SectionBlock>
-		)
-	return (
-		<EntityCard badge={<Badge tone={selectedVaultIsOwnedByAccount ? 'ok' : 'muted'}>{selectedVaultIsOwnedByAccount ? securityPoolCopy.owned : securityPoolCopy.readOnlyBadgeLabel}</Badge>} surface='flat' title={securityPoolCopy.selectedVault} variant='record'>
-			{gridContent}
-		</EntityCard>
-	)
-}
-export function getQueuedVaultOperation({ pendingOperation, selectedVaultOwner, securityVaultResult }: { pendingOperation: StagedOracleOperation | undefined; selectedVaultOwner: string; securityVaultResult: SecurityVaultSectionProps['securityVaultResult'] }) {
-	if (pendingOperation !== undefined && sameAddress(pendingOperation.targetVault, selectedVaultOwner)) {
-		if (securityVaultResult?.action === 'queueWithdrawRep' && pendingOperation.operation === 'withdrawRep') return { amount: pendingOperation.amount, isPendingSlot: true, operationId: pendingOperation.operationId } satisfies QueuedVaultOperationView
-	}
-	if (securityVaultResult?.queuedOperation === undefined) return undefined
-	if (securityVaultResult.action === 'queueWithdrawRep' && securityVaultResult.queuedOperation.operation === 'withdrawRep') return { amount: undefined, isPendingSlot: securityVaultResult.queuedOperation.isPendingSlot, operationId: securityVaultResult.queuedOperation.operationId } satisfies QueuedVaultOperationView
-	return undefined
-}
-function getQueuedVaultOperationStatus({
-	currentTimestamp,
-	currentPoolOracleManagerDetails,
-	loadingSecurityVault,
-	queuedVaultOperation,
-	securityVaultResult,
-}: {
-	currentTimestamp: bigint | undefined
-	currentPoolOracleManagerDetails: SecurityVaultSectionProps['oracleManagerDetails']
-	loadingSecurityVault: boolean
-	queuedVaultOperation: ReturnType<typeof getQueuedVaultOperation>
-	securityVaultResult: SecurityVaultSectionProps['securityVaultResult']
-}) {
-	if (securityVaultResult?.action !== 'queueWithdrawRep') return undefined
-	if (securityVaultResult.stagedExecution !== undefined) return securityVaultResult.stagedExecution.success ? 'executed' : 'failed'
-	if (queuedVaultOperation !== undefined) return queuedVaultOperation.isPendingSlot ? 'queued' : 'manual-queued'
-	if (loadingSecurityVault || currentPoolOracleManagerDetails === undefined) return 'refreshing'
-	if (isOracleManagerPriceUsable(currentPoolOracleManagerDetails, currentTimestamp)) return 'executed'
-	return 'missing'
-}
-function VaultQueuedOperationStatusCard({
-	amountLabel,
-	amountSuffix,
-	executedTitle,
-	failedTitle,
-	missingTitle,
-	missingDescription,
-	queuedTitle,
-	queuedVaultOperation,
-	manualQueuedDescription,
-	refreshingTitle,
-	refreshingDescription,
-	status,
-	successDescription,
-	errorMessage,
-	onViewStagedOperations,
-}: {
-	amountLabel: string
-	amountSuffix: string
-	errorMessage: string | undefined
-	executedTitle: string
-	failedTitle: string
-	missingDescription: string
-	missingTitle: string
-	onViewStagedOperations: (() => void) | undefined
-	queuedTitle: string
-	queuedVaultOperation: ReturnType<typeof getQueuedVaultOperation>
-	manualQueuedDescription: string
-	refreshingDescription: string
-	refreshingTitle: string
-	status: QueuedVaultOperationStatus
-	successDescription: string
-}) {
-	if (status === undefined) return undefined
-	if (status === 'queued' || status === 'manual-queued')
-		return (
-			<WarningSurface as='section' surface='flat' variant='compact'>
-				<div className='entity-card-header'>
-					<div>
-						<h4>{queuedTitle}</h4>
-					</div>
-				</div>
-				<MetricGrid>
-					<MetricField label={commonCopy.stagedOperation}>{queuedVaultOperation === undefined ? securityPoolCopy.refreshing : `#${queuedVaultOperation.operationId.toString()}`}</MetricField>
-					{queuedVaultOperation?.amount === undefined ? null : (
-						<MetricField label={amountLabel}>
-							<CurrencyValue precision='exact' value={queuedVaultOperation.amount} suffix={amountSuffix} />
-						</MetricField>
-					)}
-				</MetricGrid>
-				{status === 'manual-queued' ? <p className='detail'>{manualQueuedDescription}</p> : null}
-				{onViewStagedOperations === undefined ? undefined : (
-					<div className='actions'>
-						<button className='secondary' type='button' onClick={onViewStagedOperations}>
-							{commonCopy.viewInStagedOperations}
-						</button>
-					</div>
-				)}
-			</WarningSurface>
-		)
-	if (status === 'failed')
-		return (
-			<section className='entity-card compact flat'>
-				<div className='entity-card-header'>
-					<div>
-						<h4>{failedTitle}</h4>
-					</div>
-					<Badge tone='blocked'>{commonCopy.failed}</Badge>
-				</div>
-				<p className='detail'>{errorMessage ?? securityPoolCopy.actionRejectedDetail}</p>
-				<p className='detail'>{commonCopy.stagedOperationRetryDetail}</p>
-			</section>
-		)
-	if (status === 'executed')
-		return (
-			<section className='entity-card compact flat'>
-				<div className='entity-card-header'>
-					<div>
-						<h4>{executedTitle}</h4>
-					</div>
-					<Badge tone='ok'>{commonCopy.executed}</Badge>
-				</div>
-				<p className='detail'>{successDescription}</p>
-			</section>
-		)
-	if (status === 'missing')
-		return (
-			<WarningSurface as='section' surface='flat' variant='compact'>
-				<div className='entity-card-header'>
-					<div>
-						<h4>{missingTitle}</h4>
-					</div>
-				</div>
-				<p className='detail'>{missingDescription}</p>
-			</WarningSurface>
-		)
-	return (
-		<section className='entity-card compact flat'>
-			<div className='entity-card-header'>
-				<div>
-					<h4>{refreshingTitle}</h4>
-				</div>
-				<Badge tone='muted'>{commonCopy.refreshingWithoutEllipsis}</Badge>
-			</div>
-			<p className='detail'>{refreshingDescription}</p>
-		</section>
-	)
-}
+type VaultActionModal = 'claim-fees' | 'deposit-rep' | 'withdraw-rep' | undefined
+
 export function SecurityVaultSection({
 	accountState,
 	compactLayout = false,
