+++ ui/statoblast/ts/features/security-pools/components/VaultQueuedOperationStatusCard.tsx
@@ -1,0 +1,157 @@
+import * as commonCopy from '@zoltar/ui-core-shared/copy/common.js'
+import * as securityPoolCopy from '../../../copy/securityPool.js'
+import { Badge } from '@zoltar/ui-core-shared/components/Badge.js'
+import { CurrencyValue } from '@zoltar/ui-core-shared/components/CurrencyValue.js'
+import { MetricField } from '@zoltar/ui-core-shared/components/MetricField.js'
+import { MetricGrid } from '@zoltar/ui-core-shared/components/MetricGrid.js'
+import { WarningSurface } from '@zoltar/ui-core-shared/components/WarningSurface.js'
+import { sameAddress } from '@zoltar/ui-core-shared/lib/address.js'
+import type { StagedOracleOperation } from '@zoltar/ui-core-shared/types/contracts.js'
+import { isOracleManagerPriceUsable } from '../lib/securityVault.js'
+import type { SecurityVaultSectionProps } from '../../types.js'
+
+type QueuedVaultOperationStatus = 'executed' | 'failed' | 'manual-queued' | 'missing' | 'queued' | 'refreshing' | undefined
+type QueuedVaultOperationView = {
+	amount: bigint | undefined
+	isPendingSlot: boolean
+	operationId: bigint
+}
+
+export function getQueuedVaultOperation({ pendingOperation, selectedVaultOwner, securityVaultResult }: { pendingOperation: StagedOracleOperation | undefined; selectedVaultOwner: string; securityVaultResult: SecurityVaultSectionProps['securityVaultResult'] }) {
+	if (pendingOperation !== undefined && sameAddress(pendingOperation.targetVault, selectedVaultOwner)) {
+		if (securityVaultResult?.action === 'queueWithdrawRep' && pendingOperation.operation === 'withdrawRep') return { amount: pendingOperation.amount, isPendingSlot: true, operationId: pendingOperation.operationId } satisfies QueuedVaultOperationView
+	}
+	if (securityVaultResult?.queuedOperation === undefined) return undefined
+	if (securityVaultResult.action === 'queueWithdrawRep' && securityVaultResult.queuedOperation.operation === 'withdrawRep') return { amount: undefined, isPendingSlot: securityVaultResult.queuedOperation.isPendingSlot, operationId: securityVaultResult.queuedOperation.operationId } satisfies QueuedVaultOperationView
+	return undefined
+}
+
+export function getQueuedVaultOperationStatus({
+	currentTimestamp,
+	currentPoolOracleManagerDetails,
+	loadingSecurityVault,
+	queuedVaultOperation,
+	securityVaultResult,
+}: {
+	currentTimestamp: bigint | undefined
+	currentPoolOracleManagerDetails: SecurityVaultSectionProps['oracleManagerDetails']
+	loadingSecurityVault: boolean
+	queuedVaultOperation: ReturnType<typeof getQueuedVaultOperation>
+	securityVaultResult: SecurityVaultSectionProps['securityVaultResult']
+}) {
+	if (securityVaultResult?.action !== 'queueWithdrawRep') return undefined
+	if (securityVaultResult.stagedExecution !== undefined) return securityVaultResult.stagedExecution.success ? 'executed' : 'failed'
+	if (queuedVaultOperation !== undefined) return queuedVaultOperation.isPendingSlot ? 'queued' : 'manual-queued'
+	if (loadingSecurityVault || currentPoolOracleManagerDetails === undefined) return 'refreshing'
+	if (isOracleManagerPriceUsable(currentPoolOracleManagerDetails, currentTimestamp)) return 'executed'
+	return 'missing'
+}
+
+export function VaultQueuedOperationStatusCard({
+	amountLabel,
+	amountSuffix,
+	executedTitle,
+	failedTitle,
+	missingTitle,
+	missingDescription,
+	queuedTitle,
+	queuedVaultOperation,
+	manualQueuedDescription,
+	refreshingTitle,
+	refreshingDescription,
+	status,
+	successDescription,
+	errorMessage,
+	onViewStagedOperations,
+}: {
+	amountLabel: string
+	amountSuffix: string
+	errorMessage: string | undefined
+	executedTitle: string
+	failedTitle: string
+	missingDescription: string
+	missingTitle: string
+	onViewStagedOperations: (() => void) | undefined
+	queuedTitle: string
+	queuedVaultOperation: ReturnType<typeof getQueuedVaultOperation>
+	manualQueuedDescription: string
+	refreshingDescription: string
+	refreshingTitle: string
+	status: QueuedVaultOperationStatus
+	successDescription: string
+}) {
+	if (status === undefined) return undefined
+	if (status === 'queued' || status === 'manual-queued')
+		return (
+			<WarningSurface as='section' surface='flat' variant='compact'>
+				<div className='entity-card-header'>
+					<div>
+						<h4>{queuedTitle}</h4>
+					</div>
+				</div>
+				<MetricGrid>
+					<MetricField label={commonCopy.stagedOperation}>{queuedVaultOperation === undefined ? securityPoolCopy.refreshing : `#${queuedVaultOperation.operationId.toString()}`}</MetricField>
+					{queuedVaultOperation?.amount === undefined ? null : (
+						<MetricField label={amountLabel}>
+							<CurrencyValue precision='exact' value={queuedVaultOperation.amount} suffix={amountSuffix} />
+						</MetricField>
+					)}
+				</MetricGrid>
+				{status === 'manual-queued' ? <p className='detail'>{manualQueuedDescription}</p> : null}
+				{onViewStagedOperations === undefined ? undefined : (
+					<div className='actions'>
+						<button className='secondary' type='button' onClick={onViewStagedOperations}>
+							{commonCopy.viewInStagedOperations}
+						</button>
+					</div>
+				)}
+			</WarningSurface>
+		)
+	if (status === 'failed')
+		return (
+			<section className='entity-card compact flat'>
+				<div className='entity-card-header'>
+					<div>
+						<h4>{failedTitle}</h4>
+					</div>
+					<Badge tone='blocked'>{commonCopy.failed}</Badge>
+				</div>
+				<p className='detail'>{errorMessage ?? securityPoolCopy.actionRejectedDetail}</p>
+				<p className='detail'>{commonCopy.stagedOperationRetryDetail}</p>
+			</section>
+		)
+	if (status === 'executed')
+		return (
+			<section className='entity-card compact flat'>
+				<div className='entity-card-header'>
+					<div>
+						<h4>{executedTitle}</h4>
+					</div>
+					<Badge tone='ok'>{commonCopy.executed}</Badge>
+				</div>
+				<p className='detail'>{successDescription}</p>
+			</section>
+		)
+	if (status === 'missing')
+		return (
+			<WarningSurface as='section' surface='flat' variant='compact'>
+				<div className='entity-card-header'>
+					<div>
+						<h4>{missingTitle}</h4>
+					</div>
+				</div>
+				<p className='detail'>{missingDescription}</p>
+			</WarningSurface>
+		)
+	return (
+		<section className='entity-card compact flat'>
+			<div className='entity-card-header'>
+				<div>
+					<h4>{refreshingTitle}</h4>
+				</div>
+				<Badge tone='muted'>{commonCopy.refreshingWithoutEllipsis}</Badge>
+			</div>
+			<p className='detail'>{refreshingDescription}</p>
+		</section>
+	)
+}
