+++ ui/statoblast/ts/tests/features/security-pools/securityVaultSection.test.tsx
@@ -3,7 +3,8 @@
 import { afterEach, beforeEach, describe, expect, test } from 'bun:test'
 import { fireEvent, within } from '@zoltar/ui-core-shared/tests/testUtils/queries.js'
 import { zeroAddress } from '@zoltar/shared/ethereum'
-import { SecurityVaultSection, SelectedVaultSummarySection } from '../../../features/security-pools/components/SecurityVaultSection.js'
+import { SecurityVaultSection } from '../../../features/security-pools/components/SecurityVaultSection.js'
+import { SelectedVaultSummarySection } from '../../../features/security-pools/components/SelectedVaultSummarySection.js'
 import { evaluateSecurityPoolState } from '../../../features/security-pools/lib/securityPoolState.js'
 import type { AccountState } from '@zoltar/ui-zoltar/types/app.js'
 import type { SecurityVaultDetails } from '@zoltar/ui-core-shared/types/contracts.js'
