+++ ui/statoblast/ts/types/app.ts
@@ -1,6 +1,6 @@
 import type { ReportingOutcomeKey } from '@zoltar/ui-core-shared/types/contracts.js'
 
-export type { AccountState, WriteOperationsParameters } from '@zoltar/ui-core-shared/types/app.js'
+export type { AccountState, TransactionCancellationParameters, TransactionLifecycleParameters, WriteOperationContext, WriteOperationsParameters } from '@zoltar/ui-core-shared/types/app.js'
 export type { MarketFormState } from '@zoltar/ui-zoltar/types/app.js'
 
 export type Route = 'deploy' | 'security-pools' | 'open-oracle' | 'not-found'
