+++ ui/zoltar/ts/features/universes/hooks/useZoltarFork.ts
@@ -1,6 +1,6 @@
 import { useSignal } from '@preact/signals'
 import { useCallback, useEffect, useRef } from 'preact/hooks'
-import { zeroAddress, type Address, type Hash } from '@zoltar/shared/ethereum'
+import { zeroAddress, type Address } from '@zoltar/shared/ethereum'
 import { ABIS } from '@zoltar/ui-core-shared/abis.js'
 import { Zoltar_Zoltar } from '@zoltar/ui-core-shared/contractArtifact.js'
 import { approveErc20, forkZoltarUniverse, getZoltarAddress, readOptionalMulticall } from '../../../protocol/index.js'
@@ -17,32 +17,26 @@ import type { TokenApprovalState } from '@zoltar/ui-core-shared/lib/tokenApprova
 import { getGenesisReputationTokenAddress } from '../lib/universe.js'
 import { useRequestGuard } from '@zoltar/ui-core-shared/lib/requestGuard.js'
 import { refreshWalletStateOnly } from '@zoltar/ui-core-shared/lib/refreshState.js'
-import type { WriteOperationsParameters } from '../../../types/app.js'
+import type { TransactionLifecycleParameters, WriteOperationContext } from '../../../types/app.js'
+import type { CreateWriteClientCallbacks } from '@zoltar/ui-core-shared/lib/chainBackend.js'
 import type { ZoltarForkActionResult, ZoltarUniverseSummary } from '@zoltar/ui-core-shared/types/contracts.js'
 
-type UseZoltarForkParameters = {
-	accountAddress: Address | undefined
-	activeUniverseId: bigint
-	environmentRefreshKey: number
-	ensureZoltarUniverse: () => Promise<ZoltarUniverseSummary>
-	onTransactionFailed?: WriteOperationsParameters['onTransactionFailed']
-	onTransactionFinished: () => void
-	onTransactionPresented: WriteOperationsParameters['onTransactionPresented']
-	onTransactionPrepared?: WriteOperationsParameters['onTransactionPrepared']
-	onTransactionRequested: WriteOperationsParameters['onTransactionRequested']
-	onTransactionSubmitted: (hash: Hash) => void
-	refreshState: WriteOperationsParameters['refreshState']
-	refreshZoltarUniverse: () => Promise<ZoltarUniverseSummary | undefined>
-	shouldAutoLoadForkAccess: boolean
-	zoltarUniverse: ZoltarUniverseSummary | undefined
-}
+type UseZoltarForkParameters = TransactionLifecycleParameters &
+	WriteOperationContext & {
+		activeUniverseId: bigint
+		environmentRefreshKey: number
+		ensureZoltarUniverse: () => Promise<ZoltarUniverseSummary>
+		refreshZoltarUniverse: () => Promise<ZoltarUniverseSummary | undefined>
+		shouldAutoLoadForkAccess: boolean
+		zoltarUniverse: ZoltarUniverseSummary | undefined
+	}
 
 type OptionalReadResult<TResult> = { result: TResult; status: 'success' } | { error: Error; result?: undefined; status: 'failure' }
 type ZoltarForkAccessChildUniverse = ZoltarUniverseSummary['childUniverses'][number]
 
 export type UseZoltarForkDependencies = {
-	approveForkRep: (accountAddress: Address, callbacks: { onTransactionPrepared?: WriteOperationsParameters['onTransactionPrepared']; onTransactionSubmitted: (hash: Hash) => void }, reputationToken: Address, amount: bigint, questionId: bigint, universeId: bigint) => Promise<ZoltarForkActionResult>
-	forkZoltarUniverse: (accountAddress: Address, callbacks: { onTransactionPrepared?: WriteOperationsParameters['onTransactionPrepared']; onTransactionSubmitted: (hash: Hash) => void }, universeId: bigint, questionId: bigint) => Promise<ZoltarForkActionResult>
+	approveForkRep: (accountAddress: Address, callbacks: CreateWriteClientCallbacks, reputationToken: Address, amount: bigint, questionId: bigint, universeId: bigint) => Promise<ZoltarForkActionResult>
+	forkZoltarUniverse: (accountAddress: Address, callbacks: CreateWriteClientCallbacks, universeId: bigint, questionId: bigint) => Promise<ZoltarForkActionResult>
 	loadZoltarForkAccess: (accountAddress: Address, reputationToken: Address, universeId: bigint, childUniverses: ZoltarForkAccessChildUniverse[]) => Promise<readonly OptionalReadResult<bigint>[]>
 }
