+++ ui/zoltar/ts/features/universes/hooks/useZoltarMigration.ts
@@ -1,7 +1,7 @@
 import { useSignal } from '@preact/signals'
 import { useCallback } from 'preact/hooks'
 import { useFormState } from '@zoltar/ui-core-shared/hooks/useFormState.js'
-import type { Address, Hash } from '@zoltar/shared/ethereum'
+import type { Address } from '@zoltar/shared/ethereum'
 import { migrateInternalRepInZoltar, prepareRepForMigrationInZoltar } from '../../../protocol/index.js'
 import { createWalletWriteClient } from '@zoltar/ui-core-shared/lib/clients.js'
 import { formatRefreshErrorMessage, formatWriteErrorMessage } from '@zoltar/ui-core-shared/lib/errors.js'
@@ -14,25 +14,18 @@ import { parseBigIntListInput } from '@zoltar/ui-core-shared/lib/inputs.js'
 import { getDefaultZoltarMigrationFormState } from '../../../lib/formDefaults.js'
 import { parseRepAmountInput } from '@zoltar/ui-core-shared/lib/formInputs.js'
 import { refreshWalletStateOnly } from '@zoltar/ui-core-shared/lib/refreshState.js'
-import type { WriteOperationsParameters, ZoltarMigrationFormState } from '../../../types/app.js'
+import type { TransactionLifecycleParameters, WriteOperationContext, ZoltarMigrationFormState } from '../../../types/app.js'
 import type { ZoltarMigrationActionResult, ZoltarUniverseSummary } from '@zoltar/ui-core-shared/types/contracts.js'
 
-type UseZoltarMigrationParameters = {
-	accountAddress: Address | undefined
-	activeUniverseId: bigint
-	ensureZoltarUniverse: () => Promise<ZoltarUniverseSummary>
-	onTransactionFailed?: WriteOperationsParameters['onTransactionFailed']
-	onTransactionFinished: () => void
-	onTransactionPresented: WriteOperationsParameters['onTransactionPresented']
-	onTransactionPrepared?: WriteOperationsParameters['onTransactionPrepared']
-	onTransactionRequested: WriteOperationsParameters['onTransactionRequested']
-	onTransactionSubmitted: (hash: Hash) => void
-	refreshState: WriteOperationsParameters['refreshState']
-	refreshZoltarForkAccess: (universe?: ZoltarUniverseSummary) => Promise<void>
-	refreshZoltarUniverse: () => Promise<ZoltarUniverseSummary | undefined>
-	zoltarForkRepBalanceAttoRep: bigint | undefined
-	zoltarMigrationPreparedRepBalanceAttoRep: bigint | undefined
-}
+type UseZoltarMigrationParameters = TransactionLifecycleParameters &
+	WriteOperationContext & {
+		activeUniverseId: bigint
+		ensureZoltarUniverse: () => Promise<ZoltarUniverseSummary>
+		refreshZoltarForkAccess: (universe?: ZoltarUniverseSummary) => Promise<void>
+		refreshZoltarUniverse: () => Promise<ZoltarUniverseSummary | undefined>
+		zoltarForkRepBalanceAttoRep: bigint | undefined
+		zoltarMigrationPreparedRepBalanceAttoRep: bigint | undefined
+	}
 
 type RunZoltarMigrationActionParameters = {
 	actionName: 'prepare' | 'split'
