+++ ui/zoltar/ts/types/app.ts
@@ -1,6 +1,6 @@
 import type { MarketType, ReportingOutcomeKey } from '@zoltar/ui-core-shared/types/contracts.js'
 
-export type { AccountState, WriteOperationsParameters } from '@zoltar/ui-core-shared/types/app.js'
+export type { AccountState, TransactionCancellationParameters, TransactionLifecycleParameters, WriteOperationContext, WriteOperationsParameters } from '@zoltar/ui-core-shared/types/app.js'
 
 export type Route = 'deploy' | 'zoltar' | 'open-oracle' | 'not-found'
